% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{fpo_cols}
\alias{fpo_cols}
\title{Function to extract drsimonj colors as hex codes}
\usage{
fpo_cols(...)
}
\arguments{
\item{...}{Character names of fpo_colors}
}
\description{
Function to extract drsimonj colors as hex codes
}
