% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{fpo_pal}
\alias{fpo_pal}
\title{Return function to interpolate a drsimonj color palette}
\usage{
fpo_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in fpo_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a drsimonj color palette
}
