% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_fpo}
\alias{theme_fpo}
\title{Add Estonian Government theme to ggplot chart}
\usage{
theme_fpo(
  base_size = 14,
  base_family = "Roboto Condensed",
  lines_lwd = 0.5,
  plot_grid = TRUE,
  axis_font = base_family,
  title_size = base_size * 1.5,
  legend_size = base_size,
  bg_col = "white",
  title_font = base_family,
  title_col = "#006db5",
  base_col = "black",
  axis_lines = TRUE,
  minor_grid = ifelse(plot_grid, TRUE, FALSE),
  vert_grid = ifelse(plot_grid, TRUE, FALSE),
  ticks_type = "outer",
  horz_grid = ifelse(plot_grid, TRUE, FALSE),
  alpha_leg = 0.1,
  bord_size = 0,
  legend_bg = "white",
  strip_bg = "white",
  grid_thick = 1,
  grid_type = "solid",
  ticks_xy = "xy",
  grid_cols = c("#dad5d1", "#f2ede9")
)
}
\description{
This function allows you to add the Estonian Government theme to your ggplotgraphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
theme_fpo() + scale_color_fpo()
}
\keyword{theme_fpo}
